Don't copy paste same functions in multiple files
